fix: satisfy pkg-file-count limits for managed runtime PR
Consolidate added tests and helpers into existing files and subpackages so
backend-size and per-package file-count checks both pass on this branch.
