+++ pkg/localmodels/managed_runtime.go
@@ -1,9 +1,12 @@
 package localmodels
 
 import (
+	"fmt"
 	"strings"
 
 	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/apisurface"
+	factoryconfig "github.com/portpowered/infinite-you/pkg/config"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 )
 
@@ -121,4 +124,44 @@ func primaryModelScopedResource(aggregate catalogAggregate, factoryCfg *interfac
 		return &copied
 	}
 	return nil
+}
+
+// ManagedRuntimeReadinessForFactory returns the canonical managed-runtime readiness projection
+// for one factory dependency identity using the same catalog path for packaged and authored factories.
+func ManagedRuntimeReadinessForFactory(
+	runtimeCfg *factoryconfig.LoadedFactoryConfig,
+	modelName string,
+	opts CatalogOptions,
+) (factoryapi.ManagedRuntime, error) {
+	if runtimeCfg == nil {
+		return factoryapi.ManagedRuntime{}, fmt.Errorf("runtime config is not available")
+	}
+	catalog := BuildCatalogWithOptions(runtimeCfg, opts)
+	key := CanonicalModelName(modelName)
+	if key == "" {
+		return factoryapi.ManagedRuntime{}, fmt.Errorf("%w: empty model name", apisurface.ErrModelNotFound)
+	}
+	entry, ok := catalog[key]
+	if !ok {
+		return factoryapi.ManagedRuntime{}, fmt.Errorf("%w: %s", apisurface.ErrModelNotFound, modelName)
+	}
+	return entry.Summary.ManagedRuntime, nil
+}
+
+// EnsureManagedRuntimeReadyForInvocation classifies one managed runtime using
+// the same catalog readiness projection as discovery and inspect before
+// invocation proceeds.
+func EnsureManagedRuntimeReadyForInvocation(
+	runtimeCfg *factoryconfig.LoadedFactoryConfig,
+	modelName string,
+	opts CatalogOptions,
+) (factoryapi.ManagedRuntime, error) {
+	managed, err := ManagedRuntimeReadinessForFactory(runtimeCfg, modelName, opts)
+	if err != nil {
+		return factoryapi.ManagedRuntime{}, err
+	}
+	if invocationErr := apisurface.InvocationErrorFromManagedRuntime(managed); invocationErr != nil {
+		return managed, invocationErr
+	}
+	return managed, nil
 }
