--- pkg/localmodels/runtime_dependency.go
@@ -1,31 +1,0 @@
-package localmodels
-
-import (
-	"fmt"
-
-	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
-	"github.com/portpowered/infinite-you/pkg/apisurface"
-	factoryconfig "github.com/portpowered/infinite-you/pkg/config"
-)
-
-// ManagedRuntimeReadinessForFactory returns the canonical managed-runtime readiness projection
-// for one factory dependency identity using the same catalog path for packaged and authored factories.
-func ManagedRuntimeReadinessForFactory(
-	runtimeCfg *factoryconfig.LoadedFactoryConfig,
-	modelName string,
-	opts CatalogOptions,
-) (factoryapi.ManagedRuntime, error) {
-	if runtimeCfg == nil {
-		return factoryapi.ManagedRuntime{}, fmt.Errorf("runtime config is not available")
-	}
-	catalog := BuildCatalogWithOptions(runtimeCfg, opts)
-	key := CanonicalModelName(modelName)
-	if key == "" {
-		return factoryapi.ManagedRuntime{}, fmt.Errorf("%w: empty model name", apisurface.ErrModelNotFound)
-	}
-	entry, ok := catalog[key]
-	if !ok {
-		return factoryapi.ManagedRuntime{}, fmt.Errorf("%w: %s", apisurface.ErrModelNotFound, modelName)
-	}
-	return entry.Summary.ManagedRuntime, nil
-}
