+++ pkg/service/model_catalog.go
@@ -27,6 +27,26 @@ import (
 	"go.uber.org/zap"
 )
 
+// InvocationMetric records one emitted invocation counter together with its
+// low-cardinality dimensions.
+type InvocationMetric struct {
+	Name   string
+	Labels map[string]string
+}
+
+// InvocationMetricsRecorder receives invocation counter emissions from CLI and
+// session-runtime boundaries. Implementations should treat each call as a
+// single counter increment.
+type InvocationMetricsRecorder interface {
+	RecordInvocationMetric(InvocationMetric)
+}
+
+// ModelPullMetricsRecorder receives managed-runtime pull counter emissions.
+// Implementations should treat each call as a single counter increment.
+type ModelPullMetricsRecorder interface {
+	RecordModelPullMetric(InvocationMetric)
+}
+
 // ModelService owns direct model catalog, pull, and invocation operations.
 // FactoryService keeps these methods only as a phase-one compatibility facade.
 type ModelService interface {
