--- pkg/service/model_pull_metrics.go
@@ -1,21 +1,0 @@
-package service
-
-// InvocationMetric records one emitted invocation counter together with its
-// low-cardinality dimensions.
-type InvocationMetric struct {
-	Name   string
-	Labels map[string]string
-}
-
-// InvocationMetricsRecorder receives invocation counter emissions from CLI and
-// session-runtime boundaries. Implementations should treat each call as a
-// single counter increment.
-type InvocationMetricsRecorder interface {
-	RecordInvocationMetric(InvocationMetric)
-}
-
-// ModelPullMetricsRecorder receives managed-runtime pull counter emissions.
-// Implementations should treat each call as a single counter increment.
-type ModelPullMetricsRecorder interface {
-	RecordModelPullMetric(InvocationMetric)
-}
